Fix add raw to database

+++ classes.py
@@ -178,7 +178,7 @@ class DataBase:
                     token TEXT,
                     file_path TEXT,
                     file_name TEXT,
-                    file_duration REAL,
+                    file_duration INTEGER,
                     content_type TEXT,
                     task_id TEXT,
                     status JSONB)"""
